+++ src/Controller/Admin/CompaniesController.php
@@ -102,6 +102,12 @@ class CompaniesController extends AppController
      */
     public function delete($id = null)
     {
+        $user_session = $this->request->session()->read('Auth.User');
+        if($user_session['role'] != 'admin')
+        {
+            $this->Flash->error(__('You are not authorized to delete company.'));
+            return $this->redirect(['action' => 'index']);
+        }
         $this->request->allowMethod(['post', 'delete']);
         $company = $this->Companies->get($id);
         if ($this->Companies->delete($company)) {
